+++ xl2roefact/xl2roefact.py
@@ -1,5 +1,4 @@
 #!./.wenv_xl2roefact/bin/python3 #FIXME attn to this path if intend to move in modules/
-
 """ xl2roefact - module to create a unified class for all xl2roefact and to be used as liibrary by any other external systems
 
 Identification:
@@ -7,90 +6,81 @@ Identification:
     copyright: (c) 2023 RENWare Software Systems
     author: Petre Iordanescu (petre.iordanescu@gmail.com)
 
+Deploymens:
+    Windows: `xl2roefact.exe` portable x386 CLI application
+    Linux: `xl2roefact` executable CLI shell
+
+Specifications:
+    command general format: `xl2roefact [file(s)-to-convert] COMMAND [OPTIONS]`
+    help: `xl2roefact [COMMAND] --help`
 """
 
+
 # general libraries
+import typer
 import os
-# from pathlib import Path #NOTE see if keep / need it
+from typing_extensions import Annotated
+from pathlib import Path
+from typing import Optional
 from datetime import datetime
 from colorama import Fore, Back, Style
-
 # xl2roefact specific libraries
-import rdinv as rdinv           # status #TODO: wip
-import wrxml as wrxml           # status #FIXME: not started
-import chkxml as chkxml         # status #FIXME: not started
-import ldxml as ldxml           # status #FIXME: not started
-import chkisld as chkisld       # status #FIXME: not started
+from rdinv import rdinv           # status #TODO: wip
+from wrxml import wrxml         # status #FIXME: not yet started
+from chkxml import chkxml         # status #FIXME: not yet started
+from ldxml import ldxml          # status #FIXME: not yet started
+from chkisld import chkisld        # status #FIXME: not yet started
 
 
-class BaseProc():
+""" CLI builder section
+"""
+app_cli = typer.Typer(name="xl2roefact")
 
-    def __init__(self):
-        """set directories environment: #NOTE CURRENT directory (as returned by OS `pwd` where from application is launched)
-        """
-        self.current_directory = None
-        self.source_source_files_directoryectory = None
-        self.invoice_worksheet = None
-        #
-        self.current_directory = os.getcwd()
-        print(f"{Fore.YELLOW}DEBUG-note:{Style.RESET_ALL} class `BaseProc`, method `__init__, current directory is: {self.current_directory}") #NOTE for debug purposes
-        # set directories environment: #NOTE MODULE file directory (where module is located as efective file)
-        self.source_source_files_directoryectory = os.path.split(__file__)[0] # in all cases returns a list with 2 elements, path and file
-        print(f"{Fore.YELLOW}DEBUG-note:{Style.RESET_ALL} class `BaseProc`, method `__init__, module directory is: {self.source_source_files_directoryectory}") #NOTE for debug purposes
-
-
-    def get_directories(self):
-        """simple utility method to get outside this class what directories was detected (and will be used)
-        """
-        return self.current_directory, self.source_source_files_directoryectory
+@app_cli.command()
+def about():
+    """ short application description
+    """
+    typer.echo(f"{Fore.CYAN}xl2roefact{Style.RESET_ALL} application - convert invoice files from Excel format to JSON and XML")
+    typer.echo(f"Support: {Fore.YELLOW}www.renware.eu, petre.iordanescu@gmail.com{Style.RESET_ALL}")
+    typer.echo(f"Copyright (c) 2023 RENware Software Systems.")
+    typer.echo(f"{Fore.YELLOW}Usage:{Style.RESET_ALL}")
+    typer.echo(f"\tJust xl2json with default configs: xl2roefact xl2json")
+    typer.echo(f"\tObtain more help: xl2roefact --help")
 
 
-    def set_invoice_to(self, invoice_to_process):
-        """set current "to process" invoice (act as delayed initialize instead of sending at __init__)
-        """
-        self.crt_xl_file_to_process = invoice_to_process
-        return
+@app_cli.command()
+def settings():
+    """ display application configuration parameters and settings - subject to be changed by user
+    """
+    with open("config_settings.py") as f:
+        print(f.read())
 
 
-    def set_invoice_worksheet_to(self, invoice_worksheet):
-        """set invoice worksheet name
-        """
-        self.invoice_worksheet = invoice_worksheet
+#FIXME #NOTE @231130
+''' the function should have these options:
+    - [ ] new. `file_name`; argument, pathlib.Path, optional, default "*", validators: is file, exists, ...
+    - [ ] new. `--config_file`; option, pathlib.Path, default "config_set..." (check), validators: is file, exists, readable, ...
+''' #NOTE #FIXME
+@app_cli.command()
+def xl2json(excel_files_directory: Annotated[Path, typer.Option(
+                                                    exists=True, file_okay=False, dir_okay=True, writable=True, readable=True, resolve_path=True,
+                                                    help="directory to be used to look for Excel files"
+                                                    ),
+                                             ] = "invoice_files/"):  #TODO all args are subject of CONFIG and DOCUMENTATION
+    """ extract data from an Excel file (save data to JSON format file with the same name as original file but `.json` extension)
+    """
+    typer.echo(f"*** Component {Fore.RED}xl2roefact{Style.RESET_ALL} launched on {Fore.GREEN}{datetime.now()}{Style.RESET_ALL}")
+    typer.echo(f"{Fore.YELLOW}INFO note:{Style.RESET_ALL} directory where to search Excel files is set to: {Fore.GREEN}{excel_files_directory}{Style.RESET_ALL}")
+
+    #NOTE for test --- RENware invoice #FIXME solve it when set ARGUMENT `file_name`
+    invoice_to_process = os.path.join(excel_files_directory, "fact_RENF1004.xlsx")
+    rdinv(file_to_process=invoice_to_process)
+
+    #NOTE another test --- Kraftanlagen invoice #FIXME solve it when set ARGUMENT `file_name`
+    invoice_to_process = os.path.join(excel_files_directory, "Fact _Petrom_11017969.xlsx")
+    rdinv(file_to_process=invoice_to_process)
 
 
-    def rdinv(self):
-        """implement module `RDINV`
-        """
-        # print(f"{Fore.YELLOW}DEBUG-note:{Style.RESET_ALL} class `BaseProc`, method `rdinv`, excel file to process: {self.crt_xl_file_to_process}") #NOTE for debug purposes
-        _ = rdinv.rdinv(self.crt_xl_file_to_process, self.invoice_worksheet) #TODO return should be used but for now (test phase) is ok #FIXME - ...hereuare...wip... !!!
-
-
-
-
-
-
-#
-# #TODO_@final: this section is intended to be used for EXE CLI application
 if __name__ == "__main__":
-    print(f"\n*** Module {Fore.RED} xl2roefact {Style.RESET_ALL} launched on {Fore.GREEN}{datetime.now()}{Style.RESET_ALL}")
-
-    _invoice_processor = BaseProc()
-    _crt_directory, _source_files_directory = _invoice_processor.get_directories() #TODO subject of CONFIG
-    _excel_files_directory = "invoice_files/" #TODO subject of CONFIG
-
-
-    #FIXME for test --- RENware invoice
-    invvoice_to_process = "fact_RENF1004.xlsx"
-    _invoice_processor.set_invoice_to(os.path.join(_crt_directory, _excel_files_directory, invvoice_to_process))
-    # _invoice_processor.set_invoice_worksheet_to("FACTURA FINALA") #NOTE if not specified should open firts one...
-    _invoice_processor.rdinv()
-
-
-    #FIXME another test --- Kraftanlagen invoice
-    invvoice_to_process = "Fact _Petrom_11017969.xlsx" #NOTE original document name, next is my test for splitting
-    _invoice_processor.set_invoice_to(os.path.join(_crt_directory, _excel_files_directory, invvoice_to_process))
-    # _invoice_processor.set_invoice_worksheet_to("Factura(Invoice)") #NOTE if not specified should open firts one...
-    _invoice_processor.rdinv()
-
-
+    app_cli()